% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho_v1.R
\name{.fetch_vault_v1_by_address}
\alias{.fetch_vault_v1_by_address}
\title{Internal function to fetch vault data for a specific chain}
\usage{
.fetch_vault_v1_by_address(vault_address, chain_id)
}
\arguments{
\item{vault_address}{Character. Ethereum address of the vault}

\item{chain_id}{Integer. Chain ID}
}
\value{
A list with vault metrics or NULL if not found
}
\description{
Internal function to fetch vault data for a specific chain
}
\keyword{internal}
