% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho_v1.R
\name{.fetch_vault_v1_depositors}
\alias{.fetch_vault_v1_depositors}
\title{Internal function to fetch vault depositors for a specific chain}
\usage{
.fetch_vault_v1_depositors(vault_address, first, chain_id, verbose)
}
\arguments{
\item{vault_address}{Character. Ethereum address of the vault}

\item{first}{Integer. Number of depositors}

\item{chain_id}{Integer. Chain ID}

\item{verbose}{Logical. Print progress}
}
\value{
A tibble with depositor information
}
\description{
Internal function to fetch vault depositors for a specific chain
}
\keyword{internal}
