% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho_v1.R
\name{print.morpho_positioning_data}
\alias{print.morpho_positioning_data}
\title{Print method for morpho_positioning_data}
\usage{
\method{print}{morpho_positioning_data}(x, ...)
}
\arguments{
\item{x}{morpho_positioning_data object}

\item{...}{Additional arguments}
}
\description{
Print method for morpho_positioning_data
}
