% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_market_data}
\alias{get_market_data}
\title{Get Market Data}
\usage{
get_market_data(
  assets = c("Bitcoin", "Ethereum"),
  start_date = "2020-01-01",
  include_eth
)
}
\arguments{
\item{assets}{Character vector of assets names (Bitcoin and Ethereum are default)}

\item{start_date}{Date for the start of data collection}

\item{include_eth}{Bool}
}
\value{
List containing price data and metadata
}
\description{
Get Market Data
}
